More care around docker compose helper

+++ ci/shared/_docker_helper.sh
@@ -3,12 +3,22 @@
 function docker_compose_run() {
     USEROPT="$(id -u):$(id -g)"
     FAILED=NO
-    docker-compose build
-    docker-compose up -d
-    if ! docker-compose run --rm -u "${USEROPT}" "$@" ; then
+    if ! docker-compose build ; then
         FAILED=YES
     fi
-    docker-compose down
+    if [ "${FAILED}" == "NO" ] ; then
+        if ! docker-compose up -d ; then
+            FAILED=YES
+        fi
+        if [ "${FAILED}" == "NO" ] ; then
+            if ! docker-compose run --rm -u "${USEROPT}" "$@" ; then
+                FAILED=YES
+            fi
+        fi
+        if ! docker-compose down ; then
+            FAILED=YES
+        fi
+    fi
     if [ "${FAILED}" != "NO" ] ; then
         return 1
     fi
